Replace token pool "validate tx" logic stubs with a new "activate" action
A pool must be activated before using it. The activation step will trigger the pool
creation event to be re-sent (so that it can be used to confirm the pool), and will
also trigger the plugin to begin sending transfers from that specific pool.

This solves a number of holes with token pool creation - including validating the
pool announcement against the local plugin and ensuring transfers are not received
before the pool is fully created.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>