Replace "announce_pool" operation with "activate_pool"
The pool announcement is a definition broadcast and does not need to be tracked
by an operation (as it is sufficiently tracked by the pending/confirmed status of
the token pool and the pending/confirmed status of its linked message).

Pool activation, on the other hand, is an asynchronous operation to an external
connector, and does require an operation to track.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>