Build a predictable response body for async "InvokeContract" calls
Instead of proxying the response from ethconnect/fabconnect, successful
"InvokeContract" calls can return a predictable response body constructed
by the Contract Manager.

The request ID (which is actually the FireFly operation ID) was the only
interesting thing returned in the first place, so starting with just that
in the new response body.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>