Allow namespace to be omitted from URLs
For all namespaced URLs, add two routes - one beginning with "namespace/{ns}/"
that continues to function as before, and another without the prefix that
simply uses the default namespace. The routes are grouped into different
categories on the Swagger UI page.

This allows using shorter URLs in the common case that you want to interact
with the default namespace.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>