Remove namespace from identity manager calls
The manager is initialized for a single namespace and can assume all calls
are scoped within that namespace.

The one exception is for "network version 1", where identities may have been
registered on the legacy "ff_system" namespace. In this case, the identity
manager will query "ff_system" instead of its assigned namespace (noting that
this will only work if "ff_system" shares a database with that namespace).

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>