Do not reject messages from an unrecognized identity
Because batch downloads happen asynchronously on download manager, it's possible
that the details for an identity claim may be downloaded after other things
created by that identity.

Instead of flatly rejecting all messages from an unknown identity, leave them
in pending state so that they can be unstuck by a later rewind.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>