Replace token pool "confirmed" state with "active" bool
Token pool states were originally given similar names to message state, but
this has ended up being confusing (because the token pool creation flow
includes a message, but the token pool states have no relation to the
identically named message states).

Old terminology:
When pool definition message is "confirmed", pool is created as "pending"
When pool gets "activated" by the connector, pool is moved to "confirmed"

New terminology:
When pool definition message is "confirmed", pool is created as "not active"
When pool gets "activated" by the connector, pool is moved to "active"

My hope is that the new terminology is easier to follow, since it does not
overlap with messaging states and more clearly denotes the action of
"activating" a token pool.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>