Allow cancelling a batch that is stuck in dispatch
Batch manager status can already be queried with "/status/batchmanager". The
response includes a flag to say if a processor is blocked, along with the
ID of the flushing batch, and the last error message.

The new API "/batches/{batchid}/cancel" can now be used to cancel a batch.
This is currently only valid for batches with transaction type
"contract_invoke_pin". It will mark all messages in the batch as "cancelled",
and for private messages, will send a new batch of type "batch_pin" containing
gap fill messages. These messages have no data, but will have a CID pointing
to the original (failed) message, and will consume nonces to allow the topic
to become unblocked.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>