Dispatch gap fill messages inline
To preserve the correct ordering of nonces, gap fill batches cannot be queued
on the normal assembly loop (which might already have other messages queued).
The special batch must be created, sealed, and dispatched immediately instead
of the cancelled batch. Messages in both batches must be updated accordingly
to move them to "cancelled" or "sent".

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>